Moved internal hooks to subfolder
