+++ src/hooks/index.ts
@@ -1,7 +1,3 @@
-export * from './useFieldRegistration';
 export * from './useFormContext';
-export * from './useFullName';
-export * from './useIsUnmounted';
-export * from './useTimeout';
 export * from './useValidation';
 export * from './useField';
