+++ src/hooks/internal/useFullName.ts
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { useFormContext } from './useFormContext';
+import { useFormContext } from '../useFormContext';
 
 /**
  * Returns the full name of the field, consisting of
